requirements.txt
attrs==19.3.0
bcrypt==3.1.7
cached-property==1.5.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colored==1.4.2
cryptography==2.9.2
distro==1.5.0
docker==4.2.2
docker-compose==1.26.2
dockerpty==0.4.1
docopt==0.6.2
idna==2.10
jsonschema==3.2.0
paramiko==2.7.1
pyaml==20.4.0
pycparser==2.20
PyNaCl==1.4.0
pyrsistent==0.16.0
python-dotenv==0.14.0
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
tabulate==0.8.7
texttable==1.6.2
tqdm==4.47.0
urllib3==1.25.9
websocket-client==0.57.0
